Abstract out gRPC services
