+++ api/proto/v1/compute.proto
@@ -4,9 +4,8 @@ option go_package = "merak/";
 import "common.proto";
 
 service MerakComputeService {
-    rpc PodHandler (InternalVMPod) returns (InternalVMPod) {}
-    rpc VMHandler (InternalVMPod) returns (InternalVMPod) {}
-    rpc TestHandler (InternalVMPod) returns (InternalVMPod) {}
+    rpc ComputeHandler (InternalComputeConfigInfo) returns (InternalComputeConfigInfo) {}
+    rpc TestHandler (InternalComputeConfigInfo) returns (InternalComputeConfigInfo) {}
 }
 
 enum VMDeployType {
