+++ services/merak-compute/module.mk
@@ -3,4 +3,4 @@ module := merak-compute
 merak.compute:: build
 
 build:
-	go build -o services/merak-compute/build/merak-compute services/merak-compute/main.go
+	go build -o services/merak-compute/build/merak-compute services/merak-compute/compute.go
